Add new features to Signal, fix exp and pow in Fiber, add i_exp to utilities and change philosophy of Equalizer

+++ headers/fiber.h
@@ -2,6 +2,7 @@
 #define FIBER_H_
 
 #include "signal.h"
+#include "utility.h"
 const double planck = 6.62607004081e-34;
 const double light_speed = 299792458;
 
@@ -22,7 +23,7 @@ class Fiber {
     void setDispersionEngineering(const double& beta2);
     void setNonlinearity(const double& gamma);
     void setFiberLength(const double& length);
-    void setTotalSteps(const int& steps);
+    void setTotalSteps(const double& steps);
     void setSamplingRate(const double& rate);
 
     void propagate(Signal& field) const;
